Modluarity implementation by creating BluetoothHandler.h and BluetoothHandler.cpp then initializing an instance in main.cpp
