+++ Posadivac_5000/src/main.cpp
@@ -2,8 +2,8 @@
 #include <Adafruit_Sensor.h>
 #include <DHT.h>
 #include <NimBLEDevice.h>
+#include "BluetoothHandler.h"
 
-// Version 4 UUID generirani na: https://www.uuidgenerator.net/
 #define SERVICE_UUID        "676fa518-e4cb-4afa-aae4-f211fe532d48"
 #define CHARACTERISTIC_UUID "a08ae7a0-11e8-483d-940c-a23d81245500"
 
@@ -12,81 +12,17 @@
 
 #define SOIL_MOISTURE_PIN 35 // Analogni pin za mjerenje vlažnosti tla
 
-
-// Deklaracija globalnih BLE varijabli
-NimBLEServer* pServer = nullptr; 
-NimBLECharacteristic* pCharacteristic;
-
 //Inicijalizacija DHT senzora
 DHT dht(DHTPIN, DHTTYPE);
 
-
-// Funkcija za inicijalizaciju BLE
-void setupBLE() { 
-    NimBLEDevice::init("ESP32-Posadivac5000");
-    pServer = NimBLEDevice::createServer();
-    pServer->setCallbacks(new ServerCallbacks());
-}
-
-// Kreiranje BLE servisa i karakteristike
-void createService() {
-    NimBLEService* pService = pServer->createService(SERVICE_UUID);
-    
-    pCharacteristic = pService->createCharacteristic(
-        CHARACTERISTIC_UUID,
-        NIMBLE_PROPERTY::READ | NIMBLE_PROPERTY::NOTIFY
-    );
-
-    pCharacteristic->setValue("Čekam...");
-    pService->start(); 
-}
-
-// Funkcija za početak oglašavanja
-void startAdvertising() {
-    NimBLEAdvertising* pAdvertising = NimBLEDevice::getAdvertising();
-    pAdvertising->addServiceUUID(SERVICE_UUID);
-    pAdvertising->setName("ESP32-Posadivac5000");
-    pAdvertising->start();
-    Serial.println("Advertising započet...");
-}
-
-// Callback klasa za praćenje povezivanja uređaja
-class ServerCallbacks : public NimBLEServerCallbacks {
-    void onConnect(NimBLEServer* pServer, NimBLEConnInfo& connInfo) override {
-        Serial.printf("Uređaj spojen: %s\n", connInfo.getAddress().toString().c_str());
-    }
-
-    void onDisconnect(NimBLEServer* pServer, NimBLEConnInfo& connInfo, int reason) override {
-        Serial.printf("Uređaj odspojen, advertising reset\n");
-        NimBLEDevice::startAdvertising();
-    }
-};
-
-
-void sendSensorData(float temperatura, float vlaznost, int vlaga_tla) {
-    if (NimBLEDevice::getServer()->getConnectedCount() > 0) {
-        String data = "{";
-        data += "\"temperature\":" + String(temperatura, 1) + ",";
-        data += "\"humidity\":" + String(vlaznost, 1) + ",";
-        data += "\"soil_moisture\":" + String(vlaga_tla);
-        data += "}";
-
-        pCharacteristic->setValue(data.c_str());
-        pCharacteristic->notify();
-        Serial.println("Podaci poslani: " + data);
-    }
-}
-
+BluetoothHandler btHandler; // Inicijalizacija objekta za upravljanje BLE
 
 void setup() {
   Serial.begin(115200);
   dht.begin();
   Serial.println("DHT22 test!");
 
-  setupBLE();         // Inicijalizacija BLE
-  createService();    // Kreiranje servisa i karakteristika
-  startAdvertising(); // Početak BLE reklamiranja
-
+  btHandler.initBLE(); // Inicijalizacija BLE
 }
 
 void loop() {
@@ -102,7 +38,19 @@ void loop() {
     return;
   }
 
-  sendSensorData(temperatura_zraka, vlaznost_zraka, vlaznost_tla);
+  btHandler.sendSensorData(temperatura_zraka, vlaznost_zraka, vlaznost_tla);
+
+  Serial.print("Temperatura: ");
+  Serial.print(temperatura_zraka);
+  Serial.println(" °C");
+
+  Serial.print("Vlažnost zraka: ");
+  Serial.print(vlaznost_zraka);
+  Serial.println(" %");
+
+  Serial.print("Vlažnost tla: ");
+  Serial.print(vlaznost_tla);
+  Serial.println(" %");
 
   delay(5000);
 }
